Move some tests to the tests directory
