+++ src/hdbscan/sl.rs
@@ -858,3 +858,4 @@ mod tests {
         }
     }
 }
+
